requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
celery>=4.4.2,<4.5.0
django-filter>=2.2.0,<2.3.0
redis>=3.4.1,<3.5.0